services/call_controller/requirements.txt
pydantic>=2.0
pydantic-settings>=2.0.0
aiohttp>=3.9.5
tenacity>=8.4.1
websockets>=12.0
structlog>=24.1.0
redis>=5.0.4
python-dotenv>=1.0.0
structlog
python-json-logger
